Refactor: Move section composables to home feature package
