+++ presentation/src/main/java/com/london/presentation/feature/home/section/CarouselSection.kt
@@ -1,4 +1,4 @@
-package com.london.presentation.shared.section
+package com.london.presentation.feature.home.section
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.BorderStroke
